# refactor: faster from_pdgid