--- a/setup.cfg
+++ b/setup.cfg
@@ -58,18 +58,21 @@ where = src
 all =
     check-manifest>=0.42
     pytest
+    pytest-benchmark
     tabulate
     pandas;python_version>"3.4"
 dev =
     check-manifest>=0.42
     pytest
+    pytest-benchmark
     tabulate
     pandas;python_version>"3.4"
 test =
     pytest
+    pytest-benchmark
     pytest-cov
     tabulate
-    pandas;python_version>"3.4" and python_version<"3.10"
+    pandas;python_version>"3.4"
 
 [options.package_data]
 * = *.csv, *.fwf, *.mcd, *.py.typed
@@ -81,7 +84,9 @@ universal = 1
 testpaths =
     tests
 junit_family = xunit2
-addopts = '-Wd'
+addopts = --benchmark-disable -ra --showlocals
+filterwarnings =
+    error
 
 [flake8]
 max-complexity = 24