This should improve the performance of `from_pdgid` by using a hashed lookup instead of a linear search. Closes #348.