@eduardo-rodrigues 2021-11-02
Hi @henryiii, this seems like to totally viable and alternative option to me. But have you run some timing tests for standard workflows and also loading explicitly a table (say the particles one) and then a second one (say the nuclei) to ensure we're good? Just being maniac as he had twice issues with explosion of time to load.