@henryiii 2021-11-02
Before:
```
------------------------------------------------------------------------------------- benchmark: 2 tests -------------------------------------------------------------------------------------
Name (time in ms)                Min                 Max               Mean            StdDev             Median               IQR            Outliers       OPS            Rounds  Iterations
----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
test_load_particle_table      7.3516 (1.0)       21.5053 (1.0)       7.9696 (1.0)      1.3971 (1.0)       7.6958 (1.0)      0.1936 (1.0)          5;12  125.4763 (1.0)         108           1
test_load_nuclei             87.5544 (11.91)    116.3791 (5.41)     95.5472 (11.99)    7.9544 (5.69)     93.5607 (12.16)    3.3593 (17.35)         2;2   10.4660 (0.08)         11           1
----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
```

After:
```
------------------------------------------------------------------------------------- benchmark: 2 tests ------------------------------------------------------------------------------------
Name (time in ms)                Min                Max               Mean            StdDev             Median               IQR            Outliers       OPS            Rounds  Iterations
---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
test_load_particle_table      5.9915 (1.0)      19.5842 (1.0)       6.3464 (1.0)      1.1441 (1.0)       6.1928 (1.0)      0.1676 (1.0)          4;11  157.5703 (1.0)         146           1
test_load_nuclei             58.0820 (9.69)     81.8952 (4.18)     61.9745 (9.77)     7.6346 (6.67)     59.1297 (9.55)     1.5327 (9.14)          2;2   16.1357 (0.10)         16           1
---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
```