@henryiii 2021-11-02
I'll be fixing the third one (the lookup one) before merging. It is checking a particle near the front of the list so the performance is similar. A larger number would be more interesting. I'll probably have it check a loop over all numbers.