@henryiii 2021-11-08
Slightly nicer, and doing a bit more for the append. New from PDGID section, roughly doubles lookup performance, and increases the rest a bit.

Old:
```
------------------------------------------------------------------------------------- benchmark: 3 tests -------------------------------------------------------------------------------------
Name (time in ms)                Min                 Max               Mean            StdDev             Median               IQR            Outliers       OPS            Rounds  Iterations
----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
test_load_particle_table      7.4517 (1.0)       20.8920 (1.83)      8.1380 (1.0)      1.2724 (4.36)      7.9353 (1.0)      0.1741 (1.0)          4;19  122.8804 (1.0)         118           1
test_from_pdgid               9.6907 (1.30)      11.4370 (1.0)      10.4456 (1.28)     0.2917 (1.0)      10.4813 (1.32)     0.3427 (1.97)         27;3   95.7337 (0.78)         95           1
test_load_nuclei_append      82.6308 (11.09)    102.7947 (8.99)     87.4490 (10.75)    5.4610 (18.72)    84.9447 (10.70)    4.0507 (23.26)         1;1   11.4352 (0.09)         12           1
----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
```

New:
```
------------------------------------------------------------------------------------- benchmark: 3 tests ------------------------------------------------------------------------------------
Name (time in ms)                Min                Max               Mean            StdDev             Median               IQR            Outliers       OPS            Rounds  Iterations
---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
test_from_pdgid               4.5469 (1.0)       7.0901 (1.0)       4.8089 (1.0)      0.2603 (1.0)       4.7585 (1.0)      0.1051 (1.0)         16;19  207.9494 (1.0)         207           1
test_load_particle_table      6.2856 (1.38)     18.8554 (2.66)      6.9040 (1.44)     1.1362 (4.37)      6.7140 (1.41)     0.2250 (2.14)         5;11  144.8426 (0.70)        125           1
test_load_nuclei_append      67.6541 (14.88)    86.8385 (12.25)    71.8240 (14.94)    6.2749 (24.11)    69.1568 (14.53)    2.4218 (23.03)         2;2   13.9229 (0.07)         15           1
---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
```